Test QDateTime behavior in DST gap
Historic QDateTime behavior when being asked to create a
QDateTime in the DST gap was to interpret the given date
as if it was in the time before that gap, mapping it to a point
in time after the gap. This has changed (after Qt 5.15) with
a04411119ead3d4473e4f0ac4bceedc585977b2f .
Since then, the given date is interpreted as if it was in the
time after the gap, thus being mapped to a point in time
before the gap.
This was not caught (in dev) by Coin because machines ran in
timezone "Atlantic/Reykjavik" which does not have DST since 1967.
This patch changes tests to always run in "Europe/Oslo".
A support function and each test-case thus need to use zone-time
instead of local time.

Fixes: QTBUG-86960
Fixes: QTBUG-89208
Change-Id: Iecce5898bf9711a10e7dfc0a25e4bbeaed1c8ade
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 3d785249ba43cf4bd895ed679bac2791e0130dc5)
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>