Add qt5compat as a dependency to qtxml documentation
QXmlSimpleReader was moved to qt5compat. In the qt6 migration guide, we
would like to point to it. This patch adds qt5compat to the dependencies
of QtXml documentation in order to be able to do this.

Task-number: QTBUG-88533
Change-Id: Ifaad595862b80d31310ed594e21168cec56133a0
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
(cherry picked from commit ca5ec3d6b93b1696b8a75fdc4cac8e781df4e3a0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>