QBindable: Disallow mutation if read-only
If a QBindable is created from a computed property, it is not possible
to actually set a value or a binding. If we try to do it anyway, we'd
get a crash. Thus we now check whether the function pointer is null
before invoking it.

Task-number: QTBUG-87153
Change-Id: I5bedb9080ccf79d9b8166b80d5733d095ed76f8d
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit e236faa75f446aa3378fb013cce6598c9e076ccb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>