QInputDevice: Parent default keyboard to core app singleton
With this we can avoid leaking the QInputDevice which is created when
the platform plugin does not provide any. The onwership is solved in a
similar way as in the plugins, except that here we have no parent which
really fits, so we use QCoreApplication::instance instead.

Change-Id: I77a212fb592ba3d5a42b2ecd486763e3b4d3410e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit bc093cd294cbf48eb93b776fc0b927f0a43fa603)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>