Set new commandLink variable when initializing QCommandLinkButton
In QPushButton::initStyleOption() there is no style option feature set
to distinguish QCommandLinkButtons. The QStyleOptionButton::CommandLink
Button feature is only set during the paintEvent, but in some cases we
need to check for this feature before painting.
In 6.0 we override initStyleOption to fix this. In 5.15 initStyleOption
is not virtual so instead introduce new commandLink variable on
QPushButtonPrivate and set it to true on the QCommandLinkButton
constructor. Change the QStyleOptionButton accordingly.

Change-Id: I8831a6be7da642dcf8830812d99681213e7515dc
(adapted from commit d3ee976acab59019fdd781691f29e6f12e4dedc8)
Reviewed-by: Doris Verria <doris.verria@qt.io>